src/cavernlog.cc: Fix to build with wx 2.9.2.
git-svn-id: file:///home/survex-svn/survex/trunk@3870 4b37db11-9a0c-4f06-9ece-9ab7cdaee568

+++ src/cavernlog.cc
@@ -123,7 +123,7 @@ CavernLogWindow::OnLinkClicked(const wxHtmlLinkInfo &link)
 	size_t i = 0;
 	while ((i = cmd.find(wxT('$'), i)) != wxString::npos) {
 	    if (++i >= cmd.size()) break;
-	    switch (cmd[i]) {
+	    switch ((int)cmd[i]) {
 		case wxT('$'):
 		    cmd.erase(i, 1);
 		    break;
